Pay price Type(Long -> Integer) 변경

+++ src/main/java/com/du/forpet/domain/dto/PayResponseDto.java
@@ -13,7 +13,7 @@ public class PayResponseDto {
 
     private PayStatus status;
 
-    private Long price;
+    private Integer price;
 
     public PayResponseDto(Pay entity) {
         this.payType = entity.getPayType();
